CHANGE LOGO COMPONENT TO MENU COMPONENT
This commit replaces the Logo component import and render with a Menu component import and render. The Logo component is no longer being used, and has been replaced by a Menu component to display navigation options. A className was also added to the surrounding div to add some marginal padding.